+++ packages/zowe-explorer/__tests__/__unit__/shared/init.unit.test.ts
@@ -15,7 +15,7 @@ import * as refreshActions from "../../../src/shared/refresh";
 import * as sharedActions from "../../../src/shared/actions";
 import * as sharedExtension from "../../../src/shared/init";
 import { Profiles } from "../../../src/Profiles";
-import * as profileUtils from "../../../src/utils/ProfilesUtils";
+import * as profUtils from "../../../src/utils/ProfilesUtils";
 import * as tempFolder from "../../../src/utils/TempFolder";
 import * as zowe from "@zowe/cli";
 import { IJestIt, ITestContext, processSubscriptions, spyOnSubscriptions } from "../../__common__/testUtils";
@@ -44,18 +44,19 @@ describe("Test src/shared/extension", () => {
                 name: "zowe.updateSecureCredentials",
                 mock: [
                     { spy: jest.spyOn(globals, "setGlobalSecurityValue"), arg: [] },
-                    { spy: jest.spyOn(profileUtils, "writeOverridesFile"), arg: [] },
+                    { spy: jest.spyOn(profUtils.ProfilesUtils, "writeOverridesFile"), arg: [] },
                 ],
             },
             {
                 name: "zowe.promptCredentials",
-                mock: [{ spy: jest.spyOn(profileUtils, "promptCredentials"), arg: [test.value] }],
+                mock: [{ spy: jest.spyOn(profUtils.ProfilesUtils, "promptCredentials"), arg: [test.value] }],
             },
             {
                 name: "onDidChangeConfiguration:1",
                 mock: [
-                    { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_TEMP_FOLDER_PATH], ret: true },
-                    { spy: jest.spyOn(tempFolder, "moveTempFolder"), arg: [undefined, test.value] },
+                    { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_LOGS_FOLDER_PATH], ret: true },
+                    { spy: jest.spyOn(globals, "initLogger"), arg: [test.value] },
+                    { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_TEMP_FOLDER_PATH], ret: false },
                     { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_AUTOMATIC_PROFILE_VALIDATION], ret: false },
                     { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_TEMP_FOLDER_HIDE], ret: false },
                     { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_SECURE_CREDENTIALS_ENABLED], ret: false },
@@ -64,6 +65,18 @@ describe("Test src/shared/extension", () => {
             {
                 name: "onDidChangeConfiguration:2",
                 mock: [
+                    { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_LOGS_FOLDER_PATH], ret: false },
+                    { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_TEMP_FOLDER_PATH], ret: true },
+                    { spy: jest.spyOn(tempFolder, "moveTempFolder"), arg: [undefined, test.value] },
+                    { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_AUTOMATIC_PROFILE_VALIDATION], ret: false },
+                    { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_TEMP_FOLDER_HIDE], ret: false },
+                    { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_SECURE_CREDENTIALS_ENABLED], ret: false },
+                ],
+            },
+            {
+                name: "onDidChangeConfiguration:3",
+                mock: [
+                    { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_LOGS_FOLDER_PATH], ret: false },
                     { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_TEMP_FOLDER_PATH], ret: false },
                     { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_AUTOMATIC_PROFILE_VALIDATION], ret: true },
                     { spy: jest.spyOn(Profiles, "getInstance"), arg: [], ret: { refresh: jest.fn() } },
@@ -75,8 +88,9 @@ describe("Test src/shared/extension", () => {
                 ],
             },
             {
-                name: "onDidChangeConfiguration:3",
+                name: "onDidChangeConfiguration:4",
                 mock: [
+                    { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_LOGS_FOLDER_PATH], ret: false },
                     { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_TEMP_FOLDER_PATH], ret: false },
                     { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_AUTOMATIC_PROFILE_VALIDATION], ret: false },
                     { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_TEMP_FOLDER_HIDE], ret: true },
@@ -85,8 +99,9 @@ describe("Test src/shared/extension", () => {
                 ],
             },
             {
-                name: "onDidChangeConfiguration:4",
+                name: "onDidChangeConfiguration:5",
                 mock: [
+                    { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_LOGS_FOLDER_PATH], ret: false },
                     { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_TEMP_FOLDER_PATH], ret: false },
                     { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_AUTOMATIC_PROFILE_VALIDATION], ret: false },
                     { spy: jest.spyOn(test.value, "affectsConfiguration"), arg: [globals.SETTINGS_TEMP_FOLDER_HIDE], ret: false },