+++ packages/zowe-explorer/__tests__/__unit__/utils/LoggerUtils.unit.test.ts
@@ -9,6 +9,7 @@
  *
  */
 
+import * as path from "path";
 import * as logger from "../../../src/utils/LoggerUtils";
 import * as vscode from "vscode";
 import * as zowe from "@zowe/cli";
@@ -22,6 +23,7 @@ function createGlobalMocks() {
         mockMessage: "fake message",
         outputChannel: shared.createOutputChannel(),
         mockGetConfiguration: jest.fn(),
+        mockLogger: jest.fn(),
         testContext: {} as unknown as vscode.ExtensionContext,
     };
     newMocks.testContext = {
@@ -40,7 +42,7 @@ function createGlobalMocks() {
         configurable: true,
     });
     Object.defineProperty(Gui, "infoMessage", { value: jest.fn(), configurable: true });
-    Object.defineProperty(globals, "LOG", { value: jest.fn(), configurable: true });
+    Object.defineProperty(globals, "LOG", { value: newMocks.mockLogger, configurable: true });
     Object.defineProperty(globals.LOG, "trace", { value: jest.fn(), configurable: true });
     Object.defineProperty(globals.LOG, "debug", { value: jest.fn(), configurable: true });
     Object.defineProperty(globals.LOG, "info", { value: jest.fn(), configurable: true });
@@ -61,6 +63,19 @@ function createGlobalMocks() {
 
 describe("Logger Utils Unit Tests - function initializeZoweLogger", () => {
     const env = process.env;
+    const getSettingMock = jest.fn((section: string) => {
+        switch (section) {
+            case "cliLoggerSetting.presented":
+                return false;
+            case "files.logsFolder.path":
+                return "";
+            case "logger":
+                return "INFO";
+            default:
+                throw new Error("Unknown Configuration Setting");
+        }
+    });
+
     beforeEach(() => {
         jest.resetModules();
         process.env = { ...env };
@@ -71,19 +86,11 @@ describe("Logger Utils Unit Tests - function initializeZoweLogger", () => {
     });
     it("should initialize loggers successfully with no cli logger setting", async () => {
         const globalMocks = createGlobalMocks();
-        jest.spyOn(globals, "initLogger").mockImplementationOnce(() => {
-            return "/fake/file/path";
-        });
-        globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => "INFO",
-            update: jest.fn(),
+        jest.spyOn(globals, "initLogger").mockReturnValueOnce();
+        globalMocks.mockGetConfiguration.mockReturnValue({
+            get: getSettingMock,
         });
         const infoSpy = jest.spyOn(logger.ZoweLogger, "info");
-        process.env.ZOWE_APP_LOG_LEVEL = undefined;
-        globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => "true",
-            update: jest.fn(),
-        });
 
         expect(await logger.ZoweLogger.initializeZoweLogger(globalMocks.testContext)).toBeUndefined();
         expect(infoSpy).toHaveBeenCalled();
@@ -91,23 +98,12 @@ describe("Logger Utils Unit Tests - function initializeZoweLogger", () => {
     });
     it("should initialize loggers successfully with not changing to cli logger setting", async () => {
         const globalMocks = createGlobalMocks();
-        jest.spyOn(globals, "initLogger").mockImplementationOnce(() => {
-            return "/fake/file/path";
-        });
-        globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => "INFO",
-            update: jest.fn(),
+        jest.spyOn(globals, "initLogger").mockReturnValueOnce();
+        globalMocks.mockGetConfiguration.mockReturnValue({
+            get: getSettingMock,
         });
         const infoSpy = jest.spyOn(logger.ZoweLogger, "info");
         process.env.ZOWE_APP_LOG_LEVEL = "DEBUG";
-        globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => "INFO",
-            update: jest.fn(),
-        });
-        globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => false,
-            update: jest.fn(),
-        });
         const messageSpy = jest.spyOn(Gui, "infoMessage").mockResolvedValueOnce(undefined);
         const updateSpy = jest.spyOn(SettingsConfig, "setDirectValue");
 
@@ -122,23 +118,12 @@ describe("Logger Utils Unit Tests - function initializeZoweLogger", () => {
     });
     it("should initialize loggers successfully with changing to cli logger setting", async () => {
         const globalMocks = createGlobalMocks();
-        jest.spyOn(globals, "initLogger").mockImplementationOnce(() => {
-            return "/fake/file/path";
-        });
-        globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => "INFO",
-            update: jest.fn(),
+        jest.spyOn(globals, "initLogger").mockReturnValueOnce();
+        globalMocks.mockGetConfiguration.mockReturnValue({
+            get: getSettingMock,
         });
         const infoSpy = jest.spyOn(logger.ZoweLogger, "info");
         process.env.ZOWE_APP_LOG_LEVEL = "DEBUG";
-        globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => "INFO",
-            update: jest.fn(),
-        });
-        globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => false,
-            update: jest.fn(),
-        });
         const messageSpy = jest.spyOn(Gui, "infoMessage").mockResolvedValueOnce("Update");
         const updateSpy = jest.spyOn(SettingsConfig, "setDirectValue");
 
@@ -150,23 +135,50 @@ describe("Logger Utils Unit Tests - function initializeZoweLogger", () => {
         messageSpy.mockClear();
         updateSpy.mockClear();
     });
+    it("should reinitialize loggers successfully with new path", async () => {
+        const globalMocks = createGlobalMocks();
+        const oldLogsPath = path.join("..", "logs1");
+        const newLogsPath = path.join("..", "logs2");
+        globalMocks.mockGetConfiguration.mockReturnValue({
+            get: getSettingMock,
+        });
+        const initLoggerSpy = jest.spyOn(zowe.imperative.Logger, "initLogger").mockImplementation();
+        jest.spyOn(logger.ZoweLogger as any, "initVscLogger").mockImplementation();
+
+        await logger.ZoweLogger.initializeZoweLogger({
+            ...globalMocks.testContext,
+            extensionPath: oldLogsPath,
+        });
+        await logger.ZoweLogger.initializeZoweLogger({
+            ...globalMocks.testContext,
+            extensionPath: newLogsPath,
+        });
+
+        expect(initLoggerSpy).toHaveBeenCalledTimes(2);
+        const loggerConfig: Record<string, any> = initLoggerSpy.mock.calls[0][0];
+        expect(loggerConfig.log4jsConfig.appenders.default.filename.indexOf(oldLogsPath)).toBe(0);
+        const loggerConfig2: Record<string, any> = initLoggerSpy.mock.calls[1][0];
+        expect(loggerConfig2.log4jsConfig.appenders.default.filename.indexOf(oldLogsPath)).toBe(-1);
+        expect(loggerConfig2.log4jsConfig.appenders.default.filename.indexOf(newLogsPath)).toBe(0);
+    });
     it("should throw an error if global logger was not able to initialize", async () => {
         const globalMocks = createGlobalMocks();
         jest.spyOn(globals, "initLogger").mockImplementationOnce(() => {
             throw new Error("failed to initialize logger");
         });
-
+        globalMocks.mockLogger.mockImplementationOnce(() => {
+            throw new Error("should not call invalid logger");
+        });
         const errorMessageSpy = jest.spyOn(Gui, "errorMessage").mockImplementation();
 
         expect(await logger.ZoweLogger.initializeZoweLogger(globalMocks.testContext)).toBeUndefined();
         expect(errorMessageSpy).toBeCalledTimes(1);
+        expect(globals.LOG).not.toHaveBeenCalled();
         errorMessageSpy.mockClear();
     });
     it("should throw an error if output channel was not able to initialize", async () => {
         const globalMocks = createGlobalMocks();
-        jest.spyOn(globals, "initLogger").mockImplementationOnce(() => {
-            return "/fake/file/path";
-        });
+        jest.spyOn(globals, "initLogger").mockReturnValueOnce();
         jest.spyOn(Gui, "createOutputChannel").mockImplementationOnce(() => {
             throw new Error("failed to initialize output channel");
         });
@@ -185,8 +197,7 @@ describe("It should pass the correct message severity", () => {
     it("ZoweLogger.trace passes TRACE as severity", async () => {
         const globalMocks = createGlobalMocks();
         globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => "TRACE",
-            update: jest.fn(),
+            get: jest.fn(() => "TRACE"),
         });
         await logger.ZoweLogger.trace(globalMocks.mockMessage);
         expect(globals.LOG.trace).toHaveBeenCalled();
@@ -194,8 +205,7 @@ describe("It should pass the correct message severity", () => {
     it("ZoweLogger.debug passes DEBUG as severity", async () => {
         const globalMocks = createGlobalMocks();
         globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => "DEBUG",
-            update: jest.fn(),
+            get: jest.fn(() => "DEBUG"),
         });
         await logger.ZoweLogger.debug(globalMocks.mockMessage);
         expect(globals.LOG.debug).toHaveBeenCalled();
@@ -203,8 +213,7 @@ describe("It should pass the correct message severity", () => {
     it("ZoweLogger.info passes INFO as severity", async () => {
         const globalMocks = createGlobalMocks();
         globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => "INFO",
-            update: jest.fn(),
+            get: jest.fn(() => "INFO"),
         });
         await logger.ZoweLogger.info(globalMocks.mockMessage);
         expect(globals.LOG.info).toHaveBeenCalled();
@@ -212,8 +221,7 @@ describe("It should pass the correct message severity", () => {
     it("ZoweLogger.warn passes WARN as severity", async () => {
         const globalMocks = createGlobalMocks();
         globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => "WARN",
-            update: jest.fn(),
+            get: jest.fn(() => "WARN"),
         });
         await logger.ZoweLogger.warn(globalMocks.mockMessage);
         expect(globals.LOG.warn).toHaveBeenCalled();
@@ -221,8 +229,7 @@ describe("It should pass the correct message severity", () => {
     it("ZoweLogger.error passes ERROR as severity", async () => {
         const globalMocks = createGlobalMocks();
         globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => "ERROR",
-            update: jest.fn(),
+            get: jest.fn(() => "ERROR"),
         });
         await logger.ZoweLogger.error(globalMocks.mockMessage);
         expect(globals.LOG.error).toHaveBeenCalled();
@@ -230,8 +237,7 @@ describe("It should pass the correct message severity", () => {
     it("ZoweLogger.fatal passes FATAL as severity", async () => {
         const globalMocks = createGlobalMocks();
         globalMocks.mockGetConfiguration.mockReturnValueOnce({
-            get: (setting: string) => "FATAL",
-            update: jest.fn(),
+            get: jest.fn(() => "FATAL"),
         });
         await logger.ZoweLogger.fatal(globalMocks.mockMessage);
         expect(globals.LOG.fatal).toHaveBeenCalled();