+++ packages/zowe-explorer/src/shared/init.ts
@@ -22,9 +22,10 @@ import { TsoCommandHandler } from "../command/TsoCommandHandler";
 import { MvsCommandHandler } from "../command/MvsCommandHandler";
 import { saveFile } from "../dataset/actions";
 import { saveUSSFile } from "../uss/actions";
-import { promptCredentials, writeOverridesFile } from "../utils/ProfilesUtils";
+import { ProfilesUtils } from "../utils/ProfilesUtils";
 import { ZoweLogger } from "../utils/LoggerUtils";
 import { ZoweSaveQueue } from "../abstract/ZoweSaveQueue";
+import { SettingsConfig } from "../utils/SettingsConfig";
 
 // Set up localization
 nls.config({
@@ -72,25 +73,26 @@ export function registerCommonCommands(context: vscode.ExtensionContext, provide
     context.subscriptions.push(
         vscode.commands.registerCommand("zowe.updateSecureCredentials", async () => {
             await globals.setGlobalSecurityValue();
-            writeOverridesFile();
+            ProfilesUtils.writeOverridesFile();
         })
     );
 
     context.subscriptions.push(
         vscode.commands.registerCommand("zowe.promptCredentials", async (node: IZoweTreeNode) => {
-            await promptCredentials(node);
+            await ProfilesUtils.promptCredentials(node);
         })
     );
 
     // Register functions & event listeners
     context.subscriptions.push(
         vscode.workspace.onDidChangeConfiguration(async (e) => {
+            // If the log folder location has been changed, update current log folder preference
+            if (e.affectsConfiguration(globals.SETTINGS_LOGS_FOLDER_PATH)) {
+                await ZoweLogger.initializeZoweLogger(context);
+            }
             // If the temp folder location has been changed, update current temp folder preference
             if (e.affectsConfiguration(globals.SETTINGS_TEMP_FOLDER_PATH)) {
-                const updatedPreferencesTempPath: string = vscode.workspace
-                    .getConfiguration()
-                    /* tslint:disable:no-string-literal */
-                    .get(globals.SETTINGS_TEMP_FOLDER_PATH);
+                const updatedPreferencesTempPath: string = SettingsConfig.getDirectValue(globals.SETTINGS_TEMP_FOLDER_PATH);
                 moveTempFolder(globals.SETTINGS_TEMP_FOLDER_LOCATION, updatedPreferencesTempPath);
             }
             if (e.affectsConfiguration(globals.SETTINGS_AUTOMATIC_PROFILE_VALIDATION)) {